pyproject.toml
[tool.poetry]
packages = [
    {include = 'src'},
]
name = "noteapi"
version = "0.1.0"
description = ""
authors = ["A-V-tor <avtorca4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.114.2"}
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
greenlet = "^3.1.0"
pyjwt = "^2.9.0"
aiogram = "^3.13.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
